+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/editor-mobile-bridge
 
+## 6.11.0
+- [minor] [27189951b5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/27189951b5):
+
+  - ED-5967: added API to enable links on hybrid editor
+
+## 6.10.0
+- [minor] [30b4e99377](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/30b4e99377):
+
+  - ED-5888 Add editor dark mode
+
 ## 6.9.1
 - Updated dependencies [4a84fc40e0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4a84fc40e0):
   - @atlaskit/editor-json-transformer@4.1.12